requirements.txt
# Energy Prosumers Behavior Prediction - Requirements

# Data processing
polars==1.37.1
pandas==3.0.0
numpy==2.4.2

# Machine Learning
catboost==1.2.8
scikit-learn==1.8.0
optuna==4.7.0

# Feature engineering
holidays==0.90

# Data loading
kaggle==1.8.3
python-dotenv==1.2.1
requests==2.32.5
# Optional: for notebook development (may require Windows Long Paths enabled)
# To enable: Run as Admin: New-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Control\FileSystem" -Name "LongPathsEnabled" -Value 1 -PropertyType DWORD -Force
# jupyter>=1.0.0
